+++ openbas-api/src/main/java/io/openbas/rest/payload/service/PayloadUpdateService.java
@@ -2,9 +2,11 @@ package io.openbas.rest.payload.service;
 
 import static io.openbas.helper.StreamHelper.fromIterable;
 import static io.openbas.helper.StreamHelper.iterableToSet;
+import static io.openbas.rest.payload.PayloadUtils.validateArchitecture;
 
-import io.openbas.database.model.Payload;
+import io.openbas.database.model.*;
 import io.openbas.database.repository.AttackPatternRepository;
+import io.openbas.database.repository.DocumentRepository;
 import io.openbas.database.repository.PayloadRepository;
 import io.openbas.database.repository.TagRepository;
 import io.openbas.rest.exception.ElementNotFoundException;
@@ -13,16 +15,21 @@ import io.openbas.rest.payload.form.PayloadUpdateInput;
 import jakarta.transaction.Transactional;
 import java.time.Instant;
 import lombok.RequiredArgsConstructor;
+import org.hibernate.Hibernate;
 import org.springframework.stereotype.Service;
 
 @RequiredArgsConstructor
 @Service
 public class PayloadUpdateService {
 
+  private final PayloadUtils payloadUtils;
+
+  private final PayloadService payloadService;
+
   private final TagRepository tagRepository;
   private final AttackPatternRepository attackPatternRepository;
   private final PayloadRepository payloadRepository;
-  private final PayloadUtils payloadUtils;
+  private final DocumentRepository documentRepository;
 
   @Transactional(rollbackOn = Exception.class)
   public Payload updatePayload(String payloadId, PayloadUpdateInput input) {
@@ -32,6 +39,51 @@ public class PayloadUpdateService {
         fromIterable(attackPatternRepository.findAllById(input.getAttackPatternsIds())));
     payload.setTags(iterableToSet(tagRepository.findAllById(input.getTagIds())));
     payload.setUpdatedAt(Instant.now());
-    return payloadUtils.updatePayload(input, payload);
+    return update(input, payload);
+  }
+
+  private Payload update(PayloadUpdateInput input, Payload existingPayload) {
+    PayloadType payloadType = PayloadType.fromString(existingPayload.getType());
+    validateArchitecture(payloadType.key, input.getExecutionArch());
+
+    switch (payloadType) {
+      case COMMAND:
+        Command payloadCommand = (Command) Hibernate.unproxy(existingPayload);
+        payloadUtils.copyProperties(input, payloadCommand);
+        payloadCommand = payloadRepository.save(payloadCommand);
+        this.payloadService.updateInjectorContractsForPayload(payloadCommand);
+        return payloadCommand;
+      case EXECUTABLE:
+        Executable payloadExecutable = (Executable) Hibernate.unproxy(existingPayload);
+        payloadUtils.copyProperties(input, payloadExecutable);
+        payloadExecutable.setExecutableFile(
+            documentRepository.findById(input.getExecutableFile()).orElseThrow());
+        payloadExecutable = payloadRepository.save(payloadExecutable);
+        this.payloadService.updateInjectorContractsForPayload(payloadExecutable);
+        return payloadExecutable;
+      case FILE_DROP:
+        FileDrop payloadFileDrop = (FileDrop) Hibernate.unproxy(existingPayload);
+        payloadUtils.copyProperties(input, payloadFileDrop);
+        payloadFileDrop.setFileDropFile(
+            documentRepository.findById(input.getFileDropFile()).orElseThrow());
+        payloadFileDrop = payloadRepository.save(payloadFileDrop);
+        this.payloadService.updateInjectorContractsForPayload(payloadFileDrop);
+        return payloadFileDrop;
+      case DNS_RESOLUTION:
+        DnsResolution payloadDnsResolution = (DnsResolution) Hibernate.unproxy(existingPayload);
+        payloadUtils.copyProperties(input, payloadDnsResolution);
+        payloadDnsResolution = payloadRepository.save(payloadDnsResolution);
+        this.payloadService.updateInjectorContractsForPayload(payloadDnsResolution);
+        return payloadDnsResolution;
+      case NETWORK_TRAFFIC:
+        NetworkTraffic payloadNetworkTraffic = (NetworkTraffic) Hibernate.unproxy(existingPayload);
+        payloadUtils.copyProperties(input, payloadNetworkTraffic);
+        payloadNetworkTraffic = payloadRepository.save(payloadNetworkTraffic);
+        this.payloadService.updateInjectorContractsForPayload(payloadNetworkTraffic);
+        return payloadNetworkTraffic;
+      default:
+        throw new UnsupportedOperationException(
+            "Payload type " + existingPayload.getType() + " is not supported");
+    }
   }
 }
